TPASPKT-1564 Add error handling for file-scraper
Some of the tests currently fail due to changed mimetype and version error messages. The tests in question are not worth fixing yet since a file-scraper ticket (TPASPKT-1418) which affects mimetype and version is still in the works. The errors messages just need to be reworded in those testcases before TPASPKT-1418 gets merged.